Fix remaining issues with forward plus and fix all warnings.

+++ D3D12/PipelineState.cpp
@@ -174,7 +174,7 @@ void GraphicsPipelineState::SetLineAntialias(bool lineAntiAlias)
 	m_Desc.RasterizerState.AntialiasedLineEnable = lineAntiAlias;
 }
 
-void GraphicsPipelineState::SetDepthBias(float depthBias, float depthBiasClamp, float slopeScaledDepthBias)
+void GraphicsPipelineState::SetDepthBias(int depthBias, float depthBiasClamp, float slopeScaledDepthBias)
 {
 	m_Desc.RasterizerState.SlopeScaledDepthBias = slopeScaledDepthBias;
 	m_Desc.RasterizerState.DepthBias = depthBias;
